Move jest, typescript and webpack rules into base rules folders
The new config format doesn't allow slashes in rule names so rule names
in the style `typescript/*` were already renamed to `typescript-*`.
Lets stop nesting them in folders to reflect the new rule names.
