+++ packages/eslint-plugin/plugin.js
@@ -3,8 +3,8 @@ module.exports = {
     'binary-assignment-parens': require('./lib/rules/binary-assignment-parens'),
     'class-property-semi': require('./lib/rules/class-property-semi'),
     'images-no-direct-imports': require('./lib/rules/images-no-direct-imports'),
-    'jest-no-all-mocks-methods': require('./lib/rules/jest/no-all-mocks-methods'),
-    'jest-no-snapshots': require('./lib/rules/jest/no-snapshots'),
+    'jest-no-all-mocks-methods': require('./lib/rules/jest-no-all-mocks-methods'),
+    'jest-no-snapshots': require('./lib/rules/jest-no-snapshots'),
     'jsx-no-complex-expressions': require('./lib/rules/jsx-no-complex-expressions'),
     'jsx-no-hardcoded-content': require('./lib/rules/jsx-no-hardcoded-content'),
     'jsx-prefer-fragment-wrappers': require('./lib/rules/jsx-prefer-fragment-wrappers'),
@@ -29,9 +29,9 @@ module.exports = {
     'sinon-no-restricted-features': require('./lib/rules/sinon-no-restricted-features'),
     'sinon-prefer-meaningful-assertions': require('./lib/rules/sinon-prefer-meaningful-assertions'),
     'strict-component-boundaries': require('./lib/rules/strict-component-boundaries'),
-    'typescript-prefer-pascal-case-enums': require('./lib/rules/typescript/prefer-pascal-case-enums'),
-    'typescript-prefer-singular-enums': require('./lib/rules/typescript/prefer-singular-enums'),
-    'typescript-prefer-build-client-schema': require('./lib/rules/typescript/prefer-build-client-schema'),
-    'webpack-no-unnamed-dynamic-imports': require('./lib/rules/webpack/no-unnamed-dynamic-imports'),
+    'typescript-prefer-pascal-case-enums': require('./lib/rules/typescript-prefer-pascal-case-enums'),
+    'typescript-prefer-singular-enums': require('./lib/rules/typescript-prefer-singular-enums'),
+    'typescript-prefer-build-client-schema': require('./lib/rules/typescript-prefer-build-client-schema'),
+    'webpack-no-unnamed-dynamic-imports': require('./lib/rules/webpack-no-unnamed-dynamic-imports'),
   },
 };
