+++ packages/eslint-plugin/tests/lib/rules/typescript-prefer-pascal-case-enums.test.js
@@ -1,7 +1,7 @@
 const {FlatRuleTester: RuleTester} = require('eslint/use-at-your-own-risk');
 const typescriptParser = require('@typescript-eslint/parser');
 
-const rule = require('../../../../lib/rules/typescript/prefer-pascal-case-enums');
+const rule = require('../../../lib/rules/typescript-prefer-pascal-case-enums');
 
 const ruleTester = new RuleTester({
   languageOptions: {parser: typescriptParser},
@@ -13,7 +13,7 @@ function errorWithName(name) {
   };
 }
 
-ruleTester.run('prefer-pascal-case-enums', rule, {
+ruleTester.run('typescript-prefer-pascal-case-enums', rule, {
   valid: [
     {
       code: `enum SortOrder {MostRecent, LeastRecent, Newest, Oldest}`,
